1.API/assigment7/requirements.txt
requests
py-dotenv

